+++ src/main/java/com/productorder/controller/BuyerProductOrderManageMentController.java
@@ -3,6 +3,8 @@ package com.productorder.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -20,7 +22,6 @@ import com.productorder.model.ProductOrderVO;
 
 @Controller
 @RequestMapping("/front/buyer/productorder")
-
 public class BuyerProductOrderManageMentController {
 	
 	
@@ -37,32 +38,30 @@ public class BuyerProductOrderManageMentController {
 	
 	
 @PostMapping("cancel")
-public String buyerCancelProductOrder(@RequestParam("orderId") String orderId , Model model) {
+public String buyerCancelProductOrder(@RequestParam("orderId") String orderId , Model model, HttpSession session) {
 	
 	ProductOrderVO productOrderVO = productOrderSvc.getOneProductOrder(Integer.valueOf(orderId));
 	productOrderVO.setOrderStatus(2);
 	productOrderSvc.addProductOrder(productOrderVO);
 	
-	
-	
-	List<ProductOrderVO> list = productOrderSvc.getAll();
-	model.addAttribute("productOrderList",list);
-	model.addAttribute("success" , "-(接受訂單成功)");
+//	memberId = session.getAttribute(memberId);
+//	int memberId = 1;
+//	List<ProductOrderVO> list = productOrderSvc.findByBuyerId(memberId);
+//	model.addAttribute("productOrderList",list);
+	model.addAttribute("success" , "-(刪除訂單成功)");
 	return "redirect:/front/buyer/productorder/buyerproductorderlistall";
 
 }
 
 @PostMapping("buyerconfirm")
-public String buyerConfirmProductOrder(@RequestParam("orderId") String orderId , Model model) {
-	System.out.println(orderId+"++++++++");
+public String buyerConfirmProductOrder(@RequestParam("orderId") String orderId , Model model, HttpSession session) {
+//	System.out.println(orderId+"++++++++");
 	ProductOrderVO productOrderVO = productOrderSvc.getOneProductOrder(Integer.valueOf(orderId));
 	productOrderVO.setOrderStatus(7);
 	productOrderSvc.addProductOrder(productOrderVO);
-	
-	
-	
-	List<ProductOrderVO> list = productOrderSvc.getAll();
-	model.addAttribute("productOrderList",list);
+//	int memberId = 1;
+//	List<ProductOrderVO> list = productOrderSvc.findByBuyerId(memberId);
+//	model.addAttribute("productOrderList",list);
 	model.addAttribute("success" , "-(接受訂單成功)");
 	return "redirect:/front/buyer/productorder/buyerproductorderlistall";
 }
@@ -71,10 +70,11 @@ public String buyerConfirmProductOrder(@RequestParam("orderId") String orderId ,
 	
 
 	
-	
+//ModelAttribute//////////////////////////////////////
 	//顯示buyer所有訂單
 	@ModelAttribute("buyerProductOrderList")
-	protected List<ProductOrderVO> buyerProductOrderList (Integer buyerId , Model model){
+	protected List<ProductOrderVO> buyerProductOrderList (Integer buyerId , Model model, HttpSession session){
+//		buyerId = session.getAttribute(memberId);
 		buyerId=1;
 		List<ProductOrderVO> list = productOrderSvc.findByBuyerId(buyerId);
 		return list;
@@ -82,7 +82,8 @@ public String buyerConfirmProductOrder(@RequestParam("orderId") String orderId ,
 	
 	//顯示buyer待處理訂單
 	@ModelAttribute("buyerProductOrderPendingConfirm") 
-	protected List<ProductOrderVO> buyerProductOrderPendingConfirm(Integer buyerId, Model model) {
+	protected List<ProductOrderVO> buyerProductOrderPendingConfirm(Integer buyerId, Model model, HttpSession session) {
+//		buyerId = session.getAttribute(memberId);
 		buyerId =1;
 		List<ProductOrderVO> list = productOrderSvc.getBuyerProductOrderPendingConfirm(buyerId);
 		return list;
@@ -91,7 +92,8 @@ public String buyerConfirmProductOrder(@RequestParam("orderId") String orderId ,
 	
 	//顯示buyer顯示處理中訂單
 	@ModelAttribute("buyerProductOrderProcessing")  // for select_page.html 第97 109行用 // for listAllEmp.html 第85行用
-	protected List<ProductOrderVO> buyerProductOrderSellerProcessing(Integer buyerId, Model model) {
+	protected List<ProductOrderVO> buyerProductOrderSellerProcessing(Integer buyerId, Model model, HttpSession session) {
+//		buyerId = session.getAttribute(memberId);
 		buyerId =1;
 		List<ProductOrderVO> list = productOrderSvc.getSellerProductOrderSellerProcessing(buyerId);
 		return list;
@@ -101,7 +103,8 @@ public String buyerConfirmProductOrder(@RequestParam("orderId") String orderId ,
 	
 	//顯示buyer顯示已完成訂單
 	@ModelAttribute("buyerProductOrderCompleted") 
-	protected List<ProductOrderVO> buyerProductOrderCompleted(Integer buyerId, Model model) {
+	protected List<ProductOrderVO> buyerProductOrderCompleted(Integer buyerId, Model model, HttpSession session) {
+//		buyerId = session.getAttribute(memberId);
 		buyerId =1;
 		List<ProductOrderVO> list = productOrderSvc.getSellerProductOrderCompleted(buyerId);
 		return list;
@@ -110,7 +113,7 @@ public String buyerConfirmProductOrder(@RequestParam("orderId") String orderId ,
 	
 	//顯示buyer已取消訂單
 	@ModelAttribute("buyerProductOrderCanceled") 
-	protected List<ProductOrderVO> buyerProductOrderCanceled(Integer buyerId, Model model) {
+	protected List<ProductOrderVO> buyerProductOrderCanceled(Integer buyerId, Model model, HttpSession session) {
 		buyerId =1;
 		List<ProductOrderVO> list = productOrderSvc.getProductOrderBuyerCanceled(buyerId);
 		return list;
