+++ src/main/java/com/productorder/model/ProductOrderService.java
@@ -1,5 +1,6 @@
 package com.productorder.model;
 
+import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -9,14 +10,22 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.buyer.entity.BuyerVO;
+import com.buyer.service.BuyerService;
 import com.cart.model.CartVO;
+import com.orderdetails.model.OrderDetailsVO;
 import com.product.model.ProductService;
 import com.product.model.ProductVO;
 import com.redis.JedisUtil;
 import com.seller.entity.SellerVO;
+import com.seller.service.SellerService;
+import com.util.GenerateInvoiceNumber;
 
 import redis.clients.jedis.Jedis;
 import redis.clients.jedis.JedisPool;
@@ -34,7 +43,11 @@ public class ProductOrderService {
 	@Autowired
 	ProductService productSvc;
 	
+	@Autowired
+	SellerService sellerSvc;
 	
+	@Autowired
+	BuyerService buyerSvc;
 	
 
 	public void addProductOrder(ProductOrderVO productOrderVO) {
@@ -67,7 +80,7 @@ public class ProductOrderService {
 		return repository.findAll();
 	}
 
-/////查詢賣家訂單狀態//////////////////////////////////////////////////////
+//賣家查詢訂單狀態//////////////////////////////////////////////////////
 
 	//顯示未處理訂單
 	public List<ProductOrderVO> getSellerProductOrderPendingConfirm(Integer sellerId) {
@@ -115,7 +128,7 @@ public class ProductOrderService {
 		}
 	
 	
-/////平台查詢所有訂單//////////////////////////////////////////////////////
+//平台查詢所有訂單//////////////////////////////////////////////////////
 	public List<ProductOrderVO> getProductOrderPendingConfirm() {
 		
 		List<ProductOrderVO> allProductOrders = repository.findAll();		
@@ -199,7 +212,7 @@ public class ProductOrderService {
 		return list;
 	}
 
-/////買家查詢所有訂單 下面方法Sky新增//////////////////////////////////////////////////////	
+//買家查詢訂單狀態//////////////////////////////////////////////////////	
 	public List<ProductOrderVO> findByBuyerId(Integer buyerId) {
 		return repository.findByMemberId(buyerId);
 	}
@@ -255,15 +268,15 @@ public class ProductOrderService {
 	}
 	
 	
-////////跳轉結帳頁面用/////////////////////////////////////////////////
-	// Redis 存Key List+ DTO 作法, 獲取member 所有購物車資訊
-		public Map<String, Set<CartVO>> getAllByMemberId(String memberId) {
-
-			memberId = "8";
+		//從Redis取出買家的購物車內容
+			// Redis 存Key List+ DTO 作法, 獲取member 所有購物車資訊
+		public Map<String, Set<CartVO>> getAllByMemberId(Integer memberId) {
+			String memberIdtoString = memberId.toString();
+			
 			
 			// 設定redisKey, 取得連線
-			String redisKey = "order:"+ memberId;
-			System.out.println("redisKey= " + redisKey); //測試訊息
+			String redisKey = "order:"+ memberIdtoString;
+//			System.out.println("redisKey= " + redisKey); //測試訊息
 			JedisPool jedisPool = JedisUtil.getJedisPool();
 
 			// 本方法會用到的跨區域變數
@@ -279,7 +292,7 @@ public class ProductOrderService {
 
 				// 獲得該筆Redis 內容Keys
 				Set<String> allProductId = jedis.hkeys(redisKey);
-				System.out.println("allProuductId= " + allProductId);
+//				System.out.println("allProuductId= " + allProductId);
 
 				// 宣告要裝入回傳List<CartVO> 的參數
 				String sellerCompany;
@@ -322,5 +335,70 @@ public class ProductOrderService {
 			return cartClassfi;
 		}
 	
-////////////////////////////////////////////////////////////////	
+		//新增訂單與訂單明細	
+		public ResponseEntity<?> createOrders(ShoppingCartDTO shoppingCartDTO, Integer memberId) {
+			  //拆解DTO
+				shoppingCartDTO.getCarts().forEach(cart -> {
+		        ProductOrderVO productOrder = new ProductOrderVO();
+		        BuyerVO buyerVO = buyerSvc.getOneBuyer(memberId);
+		        SellerVO sellerVO = sellerSvc.getById(cart.getSellerId());
+		        ReceiverInfoDTO receiverInfo = cart.getReceiverInfo();
+
+		        productOrder.setBuyerVO(buyerVO);
+		        productOrder.setSellerVO(sellerVO);
+		        productOrder.setReceiver(receiverInfo.getReceiver());
+		        productOrder.setMobile(receiverInfo.getMobile());
+		        productOrder.setContactAddress(receiverInfo.getContactAddress());
+		        
+		      //用來累加訂單明細裡商品的價格
+				int actualPayment = 0;
+				int detailsTotal = 0;
+				
+			  //int orderOrigPrice = actualPay + discount; //新增優惠券功能用 (//訂單總金額 = 買家實付金額 + 平台優惠折抵)
+		      //新增訂單內容
+				long currentTimeMillis = System.currentTimeMillis();
+				Timestamp currentTimestamp = new Timestamp(currentTimeMillis);
+				Timestamp crCreateDate = currentTimestamp;
+				productOrder.setOrderTime(crCreateDate);
+				productOrder.setOrderStatus(1);
+				int invoiceNumber = GenerateInvoiceNumber.generateInvoiceNumber();
+				productOrder.setInvoiceNumber(invoiceNumber);
+				productOrder.setIsDelivered(0);
+
+		        Set<OrderDetailsVO> orderDetailsSet = new HashSet<>(); //建立訂單裡的訂單明細Set
+		        for (ProductInfoDTO productInfo : cart.getProducts()) {
+		            ProductVO productVO = productSvc.getOneProduct(productInfo.getProductId());
+		            OrderDetailsVO orderDetails = new OrderDetailsVO(); //依據每個訂單Id新增每筆訂單明細
+		            
+		          //設定訂單明細的內容
+		            orderDetails.setProductOrderVO(productOrder); // 關聯訂單
+		            orderDetails.setProductVO(productVO); //關聯商品
+		            orderDetails.setPurchaseQuantity(productInfo.getQuantity()); //設定購買數量
+		            orderDetails.setIsCommented(false);
+		            orderDetails.setIsEnable(true);
+		            detailsTotal += productVO.getPrice().intValue() * orderDetails.getPurchaseQuantity(); //訂單明細金額加總
+		            orderDetailsSet.add(orderDetails);
+		        }
+		        
+		     // 檢查是否需要增加運費
+			    int shippingFee = 0;
+			    if (detailsTotal < 999) {
+			        shippingFee = 70;
+			    }
+			    actualPayment += shippingFee;
+			    int orderOrigPrice = actualPayment + detailsTotal;
+			    productOrder.setSellerPaysShipping(0);//設定賣家負擔運費
+			    productOrder.setMemberPaysShipping(shippingFee);//設置買家負擔運費
+		        productOrder.setOrderDetailss(orderDetailsSet); // 關聯訂單明細
+		        productOrder.setActualPay(actualPayment); //設置買家實付金額
+		        productOrder.setOrderOrigPrice(orderOrigPrice); //設置訂單的總金額(訂單總金額 = 買家實付金額 + 平台優惠折抵)
+		        productOrder.setOrderDetailss(orderDetailsSet); // 訂單關聯訂單明細
+
+		        
+		        // 保存訂單及訂單明細
+		        addProductOrder(productOrder);
+		    });
+		    return ResponseEntity.ok().body(Map.of("message", "Orders created successfully."));
+		}
+		
 }
