+++ src/main/resources/templates/front-end/seller/seller-order-overall.html
@@ -150,10 +150,26 @@
 		                    <div class="col-2" th:text="${productOrderVO.orderId}"></div>
 		                    <div class="col-2 " th:text="${productOrderVO.orderTime}"></div>
 		                    <div class="col-3 " th:text="${productOrderVO.actualPay}"></div>
-		                    <div class="col-2" th:text="${productOrderVO.orderStatus}"></div>
+		                   <!--  <div class="col-2" th:text="${productOrderVO.orderStatus}"></div> -->
+										                   <div class="col-2">
+								        <span th:switch="${productOrderVO.orderStatus}">
+								            <span th:case="1">賣家待確認</span>
+								            <span th:case="2">訂單不成立（買家取消）</span>
+								            <span th:case="3">訂單不成立（賣家取消）</span>
+								            <span th:case="4">訂單成立</span>
+								            <span th:case="5">賣家理貨中</span>
+								            <span th:case="6">商品已出貨</span>
+								            <span th:case="7">訂單完成</span>
+								            <span th:case="8">付款成功</span>
+								            <span th:case="9">付款失敗</span>
+								            <span th:case="10">退貨</span>
+								            <!-- You can add more cases here if there are more statuses -->
+								            <span th:case="*">未知</span> <!-- This is the default case if none of the above match -->
+								        </span>
+								    </div>
 		                    <div class="col-3 d-flex flex-column align-items-center"> 
 		                   
-		                  		  <form method="post" th:action="@{/front/seller/productorder/confirm}">
+		                  		  <form method="post" th:action="@{/front/seller/productorder/confirm}" th:if="${productOrderVO.orderStatus == 1}">
 		                    			  <button type="submit" class="product align-self-start btn-func">
 		                      				 接受訂單
 		                      			  </button>
@@ -161,7 +177,7 @@
 								  </form> 
 		                      		
 		                      		
-		               		      <form method="post" th:action="@{/front/seller/productorder/cancel}">
+		               		      <form method="post" th:action="@{/front/seller/productorder/cancel}" th:if="${productOrderVO.orderStatus == 1}">
 		                    			  <button class="product align-self-start btn-func">
 		                    			  取消訂單
 		                    		 	</button>
@@ -211,10 +227,29 @@
                     <div class="col-2" th:text="${productOrderVO.orderId}"></div>
                     <div class="col-2 " th:text="${productOrderVO.orderTime}"></div>
                     <div class="col-3 " th:text="${productOrderVO.actualPay}"></div>
-                    <div class="col-2" th:text="${productOrderVO.orderStatus}"></div>
+                    <!-- <div class="col-2" th:text="${productOrderVO.orderStatus}"></div> -->
+                     <div class="col-2">
+								        <span th:switch="${productOrderVO.orderStatus}">
+								            <span th:case="1">賣家待確認</span>
+								            <span th:case="2">訂單不成立（買家取消）</span>
+								            <span th:case="3">訂單不成立（賣家取消）</span>
+								            <span th:case="4">訂單成立</span>
+								            <span th:case="5">賣家理貨中</span>
+								            <span th:case="6">商品已出貨</span>
+								            <span th:case="7">訂單完成</span>
+								            <span th:case="8">付款成功</span>
+								            <span th:case="9">付款失敗</span>
+								            <span th:case="10">退貨</span>
+								            <!-- You can add more cases here if there are more statuses -->
+								            <span th:case="*">未知</span> <!-- This is the default case if none of the above match -->
+								        </span>
+								    </div>
+                    
+                    
+                    
                     <div class="col-3 d-flex flex-column align-items-center"> 
                    
-                  		  <form method="post" th:action="@{/front/seller/productorder/confirm}">
+                  		  <form method="post" th:action="@{/front/seller/productorder/confirm}" th:if="${productOrderVO.orderStatus == 1}">
                     			  <button type="submit" class="product align-self-start btn-func">
                       				 接受訂單
                       			  </button>
@@ -222,7 +257,7 @@
 						  </form> 
                       		
                       		
-               		      <form method="post" th:action="@{/front/seller/productorder/cancel}">
+               		      <form method="post" th:action="@{/front/seller/productorder/cancel}" th:if="${productOrderVO.orderStatus == 1}">
                     			  <button class="product align-self-start btn-func">
                     			  取消訂單
                     		 	</button>
